move application link to components
